requirements.txt
opencv-python
mediapipe
rembg
aiogram
python-dotenv